refactor file upload / download apis
