+++ apps/client/app/api/profiles/[id]/[file]/helper.ts
@@ -1,8 +1,8 @@
 import { ProfileFile } from "../../files";
-import { GetFileDownloadResponse } from "./types";
+import { GetFileUploadResponse } from "./types";
 
-export async function getProfileFileDownload(id: number, file: ProfileFile) {
-  const response = await fetch(`/api/profiles/${id}/${file}`, { method: "GET" });
-  const { url } = (await response.json()) as GetFileDownloadResponse;
+export async function getProfileFileUpload(id: number, file: ProfileFile) {
+  const response = await fetch(`/api/profiles/${id}/${file}/upload`, { method: "PUT" });
+  const { url } = (await response.json()) as GetFileUploadResponse;
   return url;
 }
