+++ apps/client/app/api/profiles/[id]/[file]/types.ts
@@ -13,4 +13,4 @@ export const paramsSchema = z.object({
   ]),
 });
 
-export type GetFileDownloadResponse = { url: string };
+export type GetFileUploadResponse = { url: string };
