--- apps/client/app/api/profiles/[id]/[file]/upload/types.ts
@@ -1,1 +1,0 @@
-export type GetFileUploadResponse = { url: string };
