--- apps/client/app/api/projects/[id]/[file]/upload/types.ts
@@ -1,1 +1,0 @@
-export type GetFileUploadResponse = { url: string };
