+++ apps/client/app/api/publications/[id]/[file]/helper.ts
@@ -1,8 +1,8 @@
 import { PublicationFile } from "../../files";
-import { GetFileDownloadResponse } from "./types";
+import { GetFileUploadResponse } from "./types";
 
-export async function getPublicationFileDownload(id: string, file: PublicationFile) {
-  const response = await fetch(`/api/publications/${id}/${file}`, { method: "GET" });
-  const { url } = (await response.json()) as GetFileDownloadResponse;
+export async function getPublicationFileUpload(id: string, file: PublicationFile) {
+  const response = await fetch(`/api/publications/${id}/${file}/upload`, { method: "PUT" });
+  const { url } = (await response.json()) as GetFileUploadResponse;
   return url;
 }
