+++ apps/client/app/api/publications/[id]/[file]/types.ts
@@ -14,4 +14,4 @@ export const paramsSchema = z.object({
   ]),
 });
 
-export type GetFileDownloadResponse = { url: string };
+export type GetFileUploadResponse = { url: string };
