+++ apps/client/src/editor/hooks/useSave.ts
@@ -1,6 +1,6 @@
 import { useSearchParams } from "next/navigation";
 
-import { getProjectFileUpload } from "../../../app/api/projects/[id]/[file]/upload/helper";
+import { getProjectFileUpload } from "../../../app/api/projects/[id]/[file]/helper";
 import { updateProject } from "../../../app/api/projects/[id]/helper";
 import { useEditorStore } from "../store";
